Create Animal Health Diagnostics Market.md
The global Animal Health Diagnostics market was valued at US$ 3987.3 million in 2022 and is projected to reach US$ 7000.8 million by 2029, at a CAGR of 8.4% during the forecast period. The influence of COVID-19 and the Russia-Ukraine War were considered while estimating market sizes.

The USA market for Global Animal Health Diagnostics market is estimated to increase from USD million in 2022 to reach USD million by 2030, at a CAGR during the forecast period of 2023 through 2030.