Create Gaucher Disease Treatment Market.md
The global Gaucher Disease Treatment market was valued at US$ 1243 million in 2022 and is projected to reach US$ 1535.2 million by 2029, at a CAGR of 3.1% during the forecast period. The influence of COVID-19 and the Russia-Ukraine War were considered while estimating market sizes.

The USA market for Global Gaucher Disease Treatment market is estimated to increase from USD million in 2022 to reach USD million by 2030, at a CAGR during the forecast period of 2023 through 2030.