Create Vacuum Deaerators Market.md
The global Vacuum Deaerators market was valued at US$ 397.2 million in 2022 and is projected to reach US$ 476.1 million by 2029, at a CAGR of 2.6% during the forecast period. The influence of COVID-19 and the Russia-Ukraine War were considered while estimating market sizes.

A vacuum deaerator is typically used as part of a demineralizer system to remove entrained gasses like CO2 and oxygen from boiler-feed water for high-pressure boilers.