Create Polymerase Chain Reaction (PCR) for Point-of-Care (POC) Diagnostics Market.md
The global Polymerase Chain Reaction (PCR) for Point-of-Care (POC) Diagnostics market was valued at US$ 3665.8 million in 2022 and is projected to reach US$ 6128.6 million by 2029, at a CAGR of 7.6% during the forecast period. The influence of COVID-19 and the Russia-Ukraine War were considered while estimating market sizes.

The USA market for Global Polymerase Chain Reaction (PCR) for Point-of-Care (POC) Diagnostics market is estimated to increase from USD million in 2022 to reach USD million by 2030, at a CAGR during the forecast period of 2023 through 2030.