Create LED Indicator Light Market.md
The global LED Indicator Light market was valued at US$ 400 million in 2023 and is projected to reach US$ 650 million by 2030, at a CAGR of 7.2% during the forecast period. The influence of COVID-19 and the Russia-Ukraine War were considered while estimating market sizes.The U.S. Market is Estimated at $ Million in 2023, While China is Forecast to Reach $ Million.