Create Surgical Mechanic Market.md
The global Surgical Mechanic market was valued at US$ 1.79 billion in 2023 and is projected to reach US$ 2.98 billion by 2030, at a CAGR of 7.3% during the forecast period. The influence of COVID-19 and the Russia-Ukraine War were considered while estimating market sizes.The U.S. Market is Estimated at $ Million in 2023, While China is to reach $ Million.