Create Cell and Gene Therapy Drug Delivery Devices Market.md
The global Cell and Gene Therapy Drug Delivery Devices market was valued at US$ 221.49 million in 2023 and is projected to reach US$ 368.29 million by 2030, at a CAGR of 7.3% during the forecast period. The influence of COVID-19 and the Russia-Ukraine War were considered while estimating market sizes.The U.S. Market is Estimated at $ Million in 2023, While China is to reach $ Million.