Create Ambulatory Surgical Centers (ASCs) Market.md
Theglobal ambulatory surgical centers (ASCs) marketis projected to witness a CAGR of 3.5% over the forecast period. The industry is likely to expand from USD 77.1 billion in 2017 to USD 97.3 billion by 2024. The growth can be primarily attributed to the increasing demand for surgery, owing to growing incidence of accidental cases. Moreover, owing to cost effective ambulatory services, people are moving towards it, and growing demand is driven by consumer and physician choice.